Move the util classes into the respective managers.
