+++ src/main/java/dev/boostio/events/AsyncPlayerPreLogin.java
@@ -2,8 +2,8 @@ package dev.boostio.events;
 
 import dev.boostio.ChatPro;
 import dev.boostio.enums.ConfigOption;
+import dev.boostio.managers.ColorManager;
 import dev.boostio.managers.ConfigManager;
-import dev.boostio.utils.ColoringUtils;
 import dev.boostio.utils.PlayerData;
 import org.bukkit.ChatColor;
 import org.bukkit.event.EventHandler;
@@ -14,9 +14,11 @@ import java.util.UUID;
 
 public class AsyncPlayerPreLogin implements Listener {
     private final ConfigManager configManager;
+    private final ColorManager colorManager;
 
     public AsyncPlayerPreLogin(ChatPro plugin) {
         this.configManager = plugin.getConfigManager();
+        this.colorManager = plugin.getColorManager();
     }
 
     @EventHandler
@@ -26,7 +28,7 @@ public class AsyncPlayerPreLogin implements Listener {
         PlayerData playerData = new PlayerData();
         ChatPro.getInstance().getPlayerData().put(uuid, playerData);
 
-        ChatColor defaultColor = ColoringUtils.convertColor(configManager.getConfigurationOption(ConfigOption.DEFAULT_CHAT_COLOR));
+        ChatColor defaultColor = this.colorManager.convertColor(configManager.getConfigurationOption(ConfigOption.DEFAULT_CHAT_COLOR));
 
         //Set default chat color.
         ChatPro.getInstance().getPlayerData().get(uuid).setChatColorName(defaultColor);
